Business Case: Aerofit - Descriptive Statistics & Probability
About Aerofit :

Aerofit is a leading brand in the field of fitness equipment. Aerofit provides a product range including machines such as treadmills, exercise bikes, gym equipment, and fitness accessories to cater to the needs of all categories of people.

we identify the characteristics of the target audience for each type of treadmill offered by the company, to provide a better recommendation of the treadmills to the new customers and decide to investigate whether there are differences across the product with respect to customer characteristics.